Add move of initial head
