+++ packages/next/src/client/components/app-router.tsx
@@ -67,37 +67,6 @@ type AppRouterProps = Omit<
   assetPrefix: string
 }
 
-function findHeadInCache(
-  cache: CacheNode,
-  parallelRoutes: FlightRouterState[1]
-): React.ReactNode {
-  const isLastItem = Object.keys(parallelRoutes).length === 0
-  if (isLastItem) {
-    return cache.head
-  }
-  for (const key in parallelRoutes) {
-    const [segment, childParallelRoutes] = parallelRoutes[key]
-    const childSegmentMap = cache.parallelRoutes.get(key)
-    if (!childSegmentMap) {
-      continue
-    }
-
-    const cacheKey = Array.isArray(segment) ? segment[1] : segment
-
-    const cacheNode = childSegmentMap.get(cacheKey)
-    if (!cacheNode) {
-      continue
-    }
-
-    const item = findHeadInCache(cacheNode, childParallelRoutes)
-    if (item) {
-      return item
-    }
-  }
-
-  return undefined
-}
-
 function isExternalURL(url: URL) {
   return url.origin !== window.location.origin
 }
@@ -123,18 +92,15 @@ function Router({
         initialParallelRoutes,
         isServer,
         location: !isServer ? window.location : null,
+        initialHead,
       }),
-    [children, initialCanonicalUrl, initialTree]
+    [children, initialCanonicalUrl, initialTree, initialHead]
   )
   const [
     { tree, cache, prefetchCache, pushRef, focusAndScrollRef, canonicalUrl },
     dispatch,
     sync,
   ] = useReducerWithReduxDevtools(reducer, initialState)
-
-  const head = useMemo(() => {
-    return findHeadInCache(cache, tree[1])
-  }, [cache, tree])
 
   useEffect(() => {
     // Ensure initialParallelRoutes is cleaned up from memory once it's used.
@@ -360,12 +326,7 @@ function Router({
     }
   }, [onPopState])
 
-  const content = (
-    <>
-      {head || initialHead}
-      {cache.subTreeData}
-    </>
-  )
+  const content = <>{cache.subTreeData}</>
 
   return (
     <PathnameContext.Provider value={pathname}>
@@ -385,6 +346,7 @@ function Router({
                 // Root node always has `url`
                 // Provided in AppTreeContext to ensure it can be overwritten in layout-router
                 url: canonicalUrl,
+                headRenderedAboveThisLevel: false,
               }}
             >
               {HotReloader ? (
